fix(Like) : merge로 인한 충돌 해결

+++ src/main/java/org/noostak/likes/dto/DecreaseResponse.java
@@ -7,9 +7,9 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class DecreaseResponse {
-    int likes;
+    long likes;
 
-    public static DecreaseResponse of(int likes){
+    public static DecreaseResponse of(long likes){
         return new DecreaseResponse(likes);
     }
 }
